@@ -9,9 +9,11 @@
     "@types/jest": "^26.0.23",
     "@types/node": "^12.20.10",
     "@types/react": "^17.0.4",
-    "@types/react-dom": "^17.0.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "@types/react-dom": "^18.2.4",
+    "classnames": "^2.3.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.11.2",
     "react-scripts": "4.0.3",
     "typescript": "^4.2.4",
     "typescript-eslint-parser": "^22.0.0",
@@ -47,6 +49,7 @@
   "devDependencies": {
     "@mate-academy/eslint-config-react-typescript": "^0.0.4",
     "eslint": "^7.25.0",
+    "eslint-plugin-cypress": "^2.13.3",
     "gh-pages": "^3.1.0",
     "node-sass": "^5.0.0"
   }
